Add Dry.Logger API based on dispatchers
This is based on a simple idea that you can set up more than one logging
backend. This is done through 3 abstractions:

  * A logger dispatcher which can be configured with one or more backends
  * A logging backend that encapsulates how exactly logging is done
  * A logging formatter which encapsulates formatting of an individual
    log entry

I've kept the old `Logger.new` for now so that existing Hanami::Logger
will work. I'd prefer to remove this before 1.0.0 though, so that Hanami
2.0 can use the final public interface.

I'm not yet sure if we'll even need a dedicated Hanami::Logger, one of
the reasons why I'm building this stuff from scratch is to be able to
make it possible to *easily* set up custom loggers without the need to
have custom classes. Even formatters could be simple procs in most of the
cases. Various common backends can be built-in too (this can include
advanced stuff like logging to external systems etc.).

For now, this is the very foundation and planned features include:

  * A block-based DSL for configuring a multi-backend logger
  * Ability to filter out logging backends based on arbitrary rules
  * Ability to create a logging context at run-time where all matching
    backends will log entries with extra meta-data
  * A public API for implementing your own backends
  * A public API for implementing your own formatters
  * Ability to switch logging modes on-the-fly at run-time - this is
    really damn useful when you're playing with stuff in a REPL and want
    to see what's going on