Add support for conditional dispatching
You can now configure `log_if` on a per-backend basis. This is a simple
function that receives a log entry object and it's meant to return
either true or false.

You can use it like that:

```ruby
logger = Dry.Logger(:test)
  .add_backend(formatter: :string) { |backend|
    backend.log_if = -> entry { entry.info? }
  }.
  .add_backend(formatter: :string) { |backend|
    backend.log_if = -> entry { entry.error? }
  }
```

In this case the first backend will be used only for info logs, and the
second one will be used only for error logs.