Ensure logger tags and context are threadsafe
Store tags and context in a thread-local storage to ensure changes to these in one thread do not leak into another.

Default tags and context (given when initializing a logger) are frozen and duplicated per-thread.

Introduce a `Dry::Logger::ExecutionContext` module to act as an interface to thread-local storage. This could later be extracted as a general utility for use across our other gems.