Ensure logger tags and context are threadsafe (#38)
Store tags and context in thread-local storage to ensure changes to these in one thread do not leak into another.

Default tags and context (given when initializing a logger) are frozen and duplicated per-thread.

Introduce a private `Dry::Logger::ExecutionContext` module to act as an interface to thread-local storage. This could later be extracted into dry-core as a general utility for use across all our gems. At that time, we could also have it support configurable fiber- vs thread-local storage. For this first pass, however, I've kept it simple and focused on threads only, since that is what was required for the fix.